move response interop classes to their own files
